# Add serialize/deserialize for bigint to enable RVFing