--- a/modules/standard/BigInteger.chpl
+++ b/modules/standard/BigInteger.chpl
@@ -4653,7 +4653,7 @@ module BigInteger {
 
   @chpldoc.nodoc
   proc type bigint.chpl__deserialize(data) {
-    var ret: bigint();
+    var ret: bigint;
     if data.localeId == chpl_nodeID {
       mpz_set(ret.mpz, data.buff);
     } else {