--- a/modules/standard/BigInteger.chpl
+++ b/modules/standard/BigInteger.chpl
@@ -4655,7 +4655,7 @@ module BigInteger {
   proc type bigint.chpl__deserialize(data) {
     var ret: bigint;
     if data.localeId == chpl_nodeID {
-      mpz_set(ret.mpz, data.buff);
+      ret.mpz = data.buff;
     } else {
       chpl_gmp_get_mpz(ret.mpz, data.localeId, data.buff[0]);
     }