--- a/test/library/standard/BigInteger/bigIntRVF.skipif
+++ b/test/library/standard/BigInteger/bigIntRVF.skipif
@@ -1,0 +1,1 @@
+CHPL_COMM == none