--- a/test/library/standard/BigInteger/bigIntRVF.skipif
+++ b/test/library/standard/BigInteger/bigIntRVF.skipif
@@ -1,1 +1,0 @@
-CHPL_COMM == none