--- a/test/library/standard/BigInteger/nestedOn.good
+++ b/test/library/standard/BigInteger/nestedOn.good
@@ -1,2 +1,2 @@
-6
-6
+false
+false