--- a/test/optimizations/remoteValueForwarding/gmp/nestedOn.bad
+++ b/test/optimizations/remoteValueForwarding/gmp/nestedOn.bad
@@ -1,1 +1,0 @@
-nestedOn.chpl:17: error: references to remote data cannot be passed to external routines like 'mpz_sizeinbase'