--- a/test/optimizations/remoteValueForwarding/gmp/nestedOn.compopts
+++ b/test/optimizations/remoteValueForwarding/gmp/nestedOn.compopts
@@ -1,1 +1,1 @@
---remote-value-forwarding
+--remote-value-forwarding -sbigintInitThrows=true