--- a/test/studies/shootout/pidigits/bradc/GMPops.chpl
+++ b/test/studies/shootout/pidigits/bradc/GMPops.chpl
@@ -17,17 +17,6 @@ inline operator mpz_t.>=(ref x: mpz_t, ref y: mpz_t) {
 }
 
 //
-// TODO: I feel very nervous that 'rhs' is not passed by reference;
-// from past experience, I thought that all mpz_t's had to be passed
-// by reference in order to make the GMP library work correctly.
-// However, if I make it ref here, I get a complaint if I send in the
-// result of another operation like the *() below.
-//
-inline operator mpz_t.=(ref lhs: mpz_t, rhs: mpz_t) {
-  mpz_set(lhs, rhs);
-}
-
-//
 // TODO: This is not ideal due to the reliance on a local temp --
 // results in extra copies and a leak.
 //