@ronawho 2023-03-29
I think it'd be worth adding the stream test as a comm count test.

Does this resolve https://github.com/chapel-lang/chapel/issues/21889? If so it probably also resolves https://github.com/Cray/chapel-private/issues/2651 and fixes the `optimizations/remoteValueForwarding/gmp/nestedOn.chpl` future