@bmcdonald3 2023-03-29
It does fix both of those problems and resolves the future. I'll update the PR.