@bmcdonald3 2023-03-29
For some reason this is making `test/studies/shootout/pidigits/bradc/pidigits-pretty1.chpl` fail with:
```
GMPops.chpl:26: error: unsupported = overload to '1*__mpz_struct' from '1*__mpz_struct'
note: assignment overloads setting array or tuples are not currently supported
```

I don't see why the addition of the serialization functions could impact that, but it does seem to...